requirements.txt
backports.lzma==0.0.10
Pillow==5.1.0
pycrypto==2.6.1
